refactor(libprovekit): extract source-transform primitives (#1335 Phase A)
Move 12 functions + 2 structs from provekit-cli::cmd_materialize to
libprovekit::core::source_transform. Public surface for Phase B's
SiteTransformKit trait. Zero behavior change.

Moved items:
  fn concept_payload_from_line
  fn concept_payload_cid_from_line
  fn strip_comment_prefix
  fn indent_realized_source
  fn verify_payload_cid
  fn canonical_value_from_json
  fn realize_spec_from_payload
  fn capture_stub_function_block
  fn splice_realized_body_into_stub_signature
  fn extract_function_body
  fn line_starts_function_declaration
  fn leading_indent
  struct CapturedStubBlock
  struct StubSignatureAndClose

cmd_materialize.rs now re-imports via `pub(crate) use
libprovekit::core::source_transform::*;`, keeping `materialize_source_text`
unchanged. No new crate dependencies (serde_json and provekit_canonicalizer
were already direct deps of libprovekit). Error types unchanged
(`Result<_, String>`).

Verification:
  cargo build -p libprovekit -p provekit-cli: zero new warnings
    (1 pre-existing unused-variable warning in cmd_bind_migrate.rs
    is unchanged from baseline)
  cargo test -p libprovekit -p provekit-cli: pre-existing local-env
    failures only (TypeScript realize plugin requires Node
    @noble/hashes/blake3.js, identical 2 cmd_bind_migrate and 6
    cmd_materialize_integration failures on baseline and post-move)
  provekit materialize --library rust-rusqlite on
    examples/rusqlite-demo-consumer: byte-identical stdout+stderr
    vs baseline

Closes #1335.
Builds toward #1334.
